[Misc][Med] move LandingPage to pages/
