+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Container } from 'reactstrap';
 
 import Footer from './components/Footer';
-import LandingPage from './components/LandingPage';
+import LandingPage from './pages/LandingPage';
 import NavBar from './components/NavBar';
 
 function App() {
